+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `ImageFormat::magic_numbers`.
 * Add `HeadlessApp::doc_test_deadline` so spawn a timeout task in doctests.
 
 # 0.21.6